ctaphid: fix implicit integer conversion warnings
Replace int with appropriately sized types (size_t, uint8_t, uint16_t,
uint32_t, int32_t) throughout ctaphid.c and device.c to fix implicit
conversion warnings and make intent explicit.

Fix ctaphid_update_status() and device_set_status() to take uint8_t
instead of int8_t/uint32_t, matching the single-byte keepalive payload
they send on the wire.

Add ctap_hid_string() for human-readable HID command names in debug
output, and add 0x prefix to hex values in printf format strings.

+++ fido2/inc/device.h
@@ -69,7 +69,7 @@ void authenticator_write_state(AuthenticatorState *s);
  * *Optional* to compile and run, but will be required to be used for proper
  FIDO2 operation with some platforms.
 */
-void device_set_status(uint32_t status);
+void device_set_status(uint8_t status);
 
 /** Returns true if button is currently pressed.  Debouncing does not need to be
  * handled.  Should not block.
